requirements.txt
asgiref==3.5.2
Django==4.1.2
django-environ==0.9.0
django-tinymce==3.5.0
django-unixtimestampfield==0.5.0
Pillow==9.2.0
psycopg2==2.9.5
python-dotenv==0.21.0
six==1.16.0
sqlparse==0.4.3
whitenoise==6.2.0
